TestNcdfUtilities.F90 now writes COARDS-compliant output
Modified the TestNcdfUtilities.F90 test program to write
a netCDF file adhering to the COARDS standard.  This means:

(1) Index arrays having the same name as the dimension that
    is used to define them
(2) Standardized variable attribute tags
(3) Adding a time dimension (even if that dimension is only
    equal to 1)

Also added the "gamap_category" tag, which is not COARDS-
compliant, but facilitates reading of the data file
by GAMAP.

We envision that TestNcdfUtilities will be a useful
template program for those who want to write their own
netCDF applications code, and as such, it deserved a
little bit of cleanup.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>