Initial commit of ncCodeCreate script
First attempt at the ncCodeCreate script, which is similar
to ncgen in that it creates a snippet of Fortran code
with the proper calls to create a netCDF file.  The calls
are based on the NcdfUtilities routines.

At present, can define filename, global attributes,
and dimensions.  Need to define variables, variable
attributes, and index arrays.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>