Add ncCodeRead to write Fortran calls for netCDF reading
Added the ncCodeRead script, which writes a Fortran
subroutine to read a netCDF file from disk.  Analogous
to ncCodeDef and ncCodeWrite.

Made modifications in all 3 scripts for consistency.

Also added a Makefile and define.h include file that can
be used to test the output of the scripts in a Fortran
program.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>