Minor fixes in reading attributes
In each of the subroutines in the NcdfUtilieis module
m_netcdf_io_readattr.F90, we now zero the value of
"attValue" before it is used.  This makes sure that
strings will be cleared, etc.

Also corrected a minor problem in TestNcdfUtilities.F90
caused by erroneous expressions with the rc variable.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>