Now make fId a local variable in ncCodeRead
The fId variable (netCDF filehandle) needs to be declared with
INTENT(INOUT) in the perl/ncCodeDef and perl/ncCodeRead scripts.
This is because fId Is initialized in the definition routine
and needs to be passed to the writing routine as an argument.

However, in perl/ncCodeRead, we can make fId a local variable
because the netCDF file is opened and closed in the same file.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>