Now only enable netCDF compression code if the library supports it
Some netCDF-4 library builds might not have been built with compression
enabled.  We now first check the netCDF include file netcdf.inc (which
is in either the $NETCDF_INCLUDE or $NETCDF_FORTRAN_INCLUDE folder) to
see if the function "nf_def_var_deflate" is defined.  If it is, then we
set a C-preprocessor switch named NC_HAS_COMPRESSION which will activate
the code to compress the netCDF output files.  Otherwise the compression
code is left disabled. This workaround was necessary in order to avoid
compile-time errors.

Tests with the geosfp_4x5_standard simulation show a decrease in file
size from approx. 200 MB (uncompressed) to 120 MB (compressed).

Also display a message at the top of the log file indicating if this
netCDF library build supports file compression.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>