Now use a regexp to test for the compiler version being used
We test the FC environment variable for the ifort, gfortran,
pgfortran|pgi or mpifort|mpif90 compiler names. We then use this to
set the local makefile variable named THE_COMPILER.

NOTE: The former COMPILER variable is no longer used.  It does not need
to be set from the Unix environment.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>