Adding metadata for data on level edges #1
We have made several modifications so that we can define metadata
for netCDF varaiables that are placed on vertical level edges
(aka interfaces) rather than on level centers.

The "lev" index array (vertical centers) is now defined from the hyam
and hybm coefficients.  Similarly, the ilev index array is defined from
the hyai and hybi coefficients.  We had to modify routines in ncdf_mod.F90
to pass a couple of additional variable attribut arguments, as well
as to get the dimension ID for "ilev".

Added the "ilev" dimension ID to the HistItem and HistContainer types.
Also, now the History Item's data array is dimensioned to the same
size as the relevant source data pointer.

This commit causes an netcdf "start+count exceeds dimension bound" but
we will resolve that in the next commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>