Turn on netCDF chunking (if file compression is supported)
NcdfUtil/ncdf_mod.F90: Added routine Nc_Var_Chunk, which turns on
netCDF chunking in a variable.

NOTE: NetCDF-4 chunking will be turned off by default if the netCDF-4
library was not built with support for compression.  It will also be
turned off if GEOS-Chem is compiled with DEBUG=y or NC_NODEFLATE=y.
(Compression can make it difficult to directly compare netCDF files
as one would when performing a difference test, so we disable it
for debugging purposes).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>