Add some useful updates to the netCDF utility modules
NcdfUtil/ncdf_mod.F90: Add number of updates as an optional global
attribute (which is currently not saved out by default).
NOTE: This is based on the "heartbeat" timestep and thus might not
be accurate if a diagnostic is only updated every chemistry timestep.

NcdfUtil/m_do_err_out.F90: Make a more readable error format.
We have added an additional sentence to the netCDF error output
to indicate that the error message was caused in either writing
to or reading from a netCDF file.

Other modules:  Use 512 characters for the error message, so that
filenames won't be truncated prematurely.

NOTE: In a future commit, we will return the error codes up to the
main program level.  Don't have time to work on it now.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>