Strip null character from VarUnit in ncdf_mod.F90
It was discovered that in many netCDF files, the ASCII null string
("\0", ASCII value = 0) is often appended to the end of the units string,
especially if the netCDF file was written by code in a language other
than Fortran.  But the Fortran compiler can interpret the null string
as part of the text instead of as a character to be skipped.

For example, when reading the unit string from a netCDF data file
via HEMCO (which uses ncdf_mod.F90), debug printout reveals:

     ThisUnit: g/m2/yr^@ : 8 characters
     OrigUnit: g/m2/yr   : 7 characters

where the ^@ is the representation of the null string.  If the unit
strings do not match exactly, this will generate a bunch of
"HEMCO WARNING: File units do not match" warnings in the HEMCO.log file.

The simple fix is to test if the last character of the units string
is the ASCII null string (in routine NC_READ_ARR).  If so, then we
immediately replace it with a Fortran empty character.  This produces
the expected behavior.  This is done in routines NC_READ_VAR_CORE
and NC_READ_ARR in NcdfUtil/ncdf_mod.F90.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>